Separate the DB connection class
